Move to own maven module
