+++ spoon-javadoc/src/main/java/spoon/javadoc/api/JavadocTagType.java
@@ -5,7 +5,7 @@
  *
  * Spoon is available either under the terms of the MIT License (see LICENSE-MIT.txt) of the Cecill-C License (see LICENSE-CECILL-C.txt). You as the user are entitled to choose the terms under which to adopt Spoon.
  */
-package spoon.javadoc.external;
+package spoon.javadoc.api;
 
 import java.util.Collection;
 import java.util.List;
