+++ spoon-javadoc/src/main/java/spoon/javadoc/api/elements/JavadocInlineTag.java
@@ -5,9 +5,9 @@
  *
  * Spoon is available either under the terms of the MIT License (see LICENSE-MIT.txt) of the Cecill-C License (see LICENSE-CECILL-C.txt). You as the user are entitled to choose the terms under which to adopt Spoon.
  */
-package spoon.javadoc.external.elements;
+package spoon.javadoc.api.elements;
 
-import spoon.javadoc.external.JavadocTagType;
+import spoon.javadoc.api.JavadocTagType;
 
 import java.util.List;
