+++ spoon-javadoc/src/main/java/spoon/javadoc/api/elements/snippets/JavadocSnippetBody.java
@@ -5,9 +5,9 @@
  *
  * Spoon is available either under the terms of the MIT License (see LICENSE-MIT.txt) of the Cecill-C License (see LICENSE-CECILL-C.txt). You as the user are entitled to choose the terms under which to adopt Spoon.
  */
-package spoon.javadoc.external.elements.snippets;
+package spoon.javadoc.api.elements.snippets;
 
-import spoon.javadoc.external.parsing.SnippetFileParser;
+import spoon.javadoc.api.parsing.SnippetFileParser;
 
 import java.util.Collections;
 import java.util.List;
